# Refactor draw pixel and hardcode another path