--- a/GPU/GPU.vcxproj
+++ b/GPU/GPU.vcxproj
@@ -453,6 +453,7 @@
     <ClInclude Include="GPUState.h" />
     <ClInclude Include="Math3D.h" />
     <ClInclude Include="Software\Clipper.h" />
+    <ClInclude Include="Software\DrawPixel.h" />
     <ClInclude Include="Software\Lighting.h" />
     <ClInclude Include="Software\FuncId.h" />
     <ClInclude Include="Software\Rasterizer.h" />
@@ -629,6 +630,7 @@
     <ClCompile Include="GPUState.cpp" />
     <ClCompile Include="Math3D.cpp" />
     <ClCompile Include="Software\Clipper.cpp" />
+    <ClCompile Include="Software\DrawPixel.cpp" />
     <ClCompile Include="Software\Lighting.cpp" />
     <ClCompile Include="Software\FuncId.cpp" />
     <ClCompile Include="Software\Rasterizer.cpp" />