--- a/GPU/GPU.vcxproj.filters
+++ b/GPU/GPU.vcxproj.filters
@@ -267,6 +267,9 @@
     <ClInclude Include="Software\FuncId.h">
       <Filter>Software</Filter>
     </ClInclude>
+    <ClInclude Include="Software\DrawPixel.h">
+      <Filter>Software</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Math3D.cpp">
@@ -534,6 +537,9 @@
       <Filter>Common</Filter>
     </ClCompile>
     <ClCompile Include="Software\FuncId.cpp">
+      <Filter>Software</Filter>
+    </ClCompile>
+    <ClCompile Include="Software\DrawPixel.cpp">
       <Filter>Software</Filter>
     </ClCompile>
   </ItemGroup>