--- a/GPU/Software/FuncId.h
+++ b/GPU/Software/FuncId.h
@@ -126,3 +126,4 @@ struct hash<SamplerID> {
 };
 
 void ComputePixelFuncID(PixelFuncID *id);
+std::string DescribePixelFuncID(const PixelFuncID &id);