… CMakeLists.txt
… GPU/Software/Rasterizer.cpp
… GPU/Software/SoftGpu.cpp
… UWP/GPU_UWP/GPU_UWP.vcxproj
… UWP/GPU_UWP/GPU_UWP.vcxproj.filters
… libretro/Makefile.common