This refactors the pixel drawing of the software renderer to a separate file, to more cleanly separate the part covered by PixelFuncID and the future pixel drawing jit.

Since a function pointer is now used for drawing pixels (flat perf impact, actually, was slightly worried), I went ahead and templated the framebuffer format.  This resulted in a small 5-10% perf improvement.  For example, Valkyria Chronicles 3's intro went from 35 FPS to 38 FPS, but some other areas only went from 105 FPS to 110 FPS.

Of course, a jitted pixel func can do better, but this is already a real improvement from just switching to a function pointer.

The pixelID parameter is a bit unfortunate... I think I'll have to keep it because I'd always want a "DISABLE", if you will.  My previous experiments were using __vectorcall, but I'd also messed with drawing four pixels at once which had its own tradeoffs.

-[Unknown]