@hrydgard 2021-11-21
That's cool. Modern branch predictors handle function pointers well as long as they predict well, which will of course be the case here.