@hrydgard 2021-11-21
There's a minor build error, the usual whether to include <string> or not, it seems...