@unknownbrackets 2021-11-21
> That's cool. Modern branch predictors handle function pointers well as long as they predict well, which will of course be the case here.

Yeah, it makes sense since it's a common need (i.e. virtual functions.)

-[Unknown]